Minor fixes, cleaned up filetree, merged sound code
Just a couple fixes for the camera, cleaned up the filetree, and merged the sound and event files in since I had them while cleaning up.

Only other thing of note is that I changed the renderer_create call in Game.cpp to use linear scaling and the GPU, we can go ahead and change this later again if it's causing problems.